gmsol_programs::gmsol_store

Module types

Program type definitions.

Note that account and event type definitions are not included in this module, as they have their own dedicated modules.

Structs

ActionFlagContainer

Flags container generated by the macro.

ActionHeader

Action Header.

Addresses

Addresses.

Amounts

Amounts.

BorrowingFees

Borrowing Fee.

ClaimableCollateral

Claimable collateral amounts.

Clocks

Market clocks.

CreateDepositParams

Create Deposit Params.

CreateGlvDepositParams

Create GLV Deposit Params.

CreateGlvWithdrawalParams

Create GLV Withdrawal Params.

CreateOrderParams

Create Order Arguments.

CreateShiftParams

Create Shift Params.

CreateWithdrawalParams

Create Withdrawal Params.

DecreasePositionReport

Report of the execution of position decreasing.

DepositActionParams

Deposit Params.

DepositParams

Deposit params.

DepositReport

Report of the execution of deposit.

DepositTokenAccounts

Token Accounts.

DisabledFeatures

Disabled Features State.

DisabledMap

Fixed size map generated by the macro.

DisabledMapEntry

Entry.

DistributePositionImpactReport

Distribute Position Impact Report.

Entry

An entry of the config buffer.

EntryArgs

An entry of the config buffer.

EventClocks

Market clocks.

EventOtherState

Market State.

EventPool

A pool for market.

EventPositionState

Position State.

EventTradeFees

Trade Fees.

EventTradeOutputAmounts

Output amounts.

EventTradePnl

Trade PnL.

EventTradePrice

Price.

EventTradePrices

Trade Prices.

EventTransferOut

Transfer Out.

ExecutionParams

Execution Params for increasing position.

Factors

Factors.

Fees

Fees.

FundingFees

Funding Fees.

GlvDepositActionParams

GLV Deposit Params.

GlvDepositTokenAccounts

Token and accounts.

GlvMarketConfig

Market Config for GLV.

GlvMarketFlagContainer

Flags container generated by the macro.

GlvMarkets

Fixed size map generated by the macro.

GlvMarketsEntry

Entry.

GlvWithdrawalActionParams

GLV Withdrawal Params.

GlvWithdrawalTokenAccounts

Token and accounts.

GtExchangeFlagContainer

Flags container generated by the macro.

GtExchangeVaultFlagContainer

Flags container generated by the macro.

GtState

IncreasePositionParams

Increase Position Params.

IncreasePositionReport

Report of the execution of position increasing.

Indexer

Market indexer.

LiquidationFees

Liquidation Fees.

MarketConfig

Market Config.

MarketConfigFlagContainer

Flags container generated by the macro.

MarketFlagContainer

Flags container generated by the macro.

MarketMeta

Market Metadata.

MarketStatus

Market Status.

Members

Fixed size map generated by the macro.

MembersEntry

Entry.

OracleFlagContainer

Flags container generated by the macro.

OrderActionParams

Order params.

OrderFees

Order Fees.

OrderTokenAccounts

Token accounts for Order.

OtherState

Market State.

OutputAmounts

Output amounts.

Pnl

Processed PnL.

Pool

A pool for market.

PoolStorage

A pool storage for market.

Pools

Market Pools.

PositionFees

Position Fees.

PositionState

Position State.

Price

Price.

PriceFeedPrice

Price structure for Price Feed.

PriceFlagContainer

Flags container generated by the macro.

PriceMap

Fixed size map generated by the macro.

PriceMapEntry

Entry.

Prices

Prices for execution.

Referral

Referral.

RevertibleBuffer

RoleMap

Fixed size map generated by the macro.

RoleMapEntry

Entry.

RoleMetadata

Role Metadata.

RoleStore

Roles Store.

ShiftActionParams

ShiftTokenAccounts

SmallPrices

Zero-copy price structure for storing min max prices.

State

SwapActionParams

Swap params.

SwapParams

Swap params.

SwapReport

Report of the execution of swap.

SwapResult

TokenAndAccount

Token Account.

Tokens

Fixed size map generated by the macro.

TokensEntry

Entry.

TradeFees

Trade Fees.

TradeOutputAmounts

Output amounts.

TradePnl

Trade PnL.

TradePrice

Price.

TradePrices

Prices.

TransferOut

Transfer Out.

Treasury

Treasury.

UpdateBorrowingReport

Update Borrowing Report.

UpdateFundingReport

Update Funding Report.

UpdateGlvParams

GLV Update Params.

UpdateOrderParams

Update Order Params.

UpdateTokenConfigParams

UserFlagContainer

Flags container generated by the macro.

UserGtState

GT State.

WithdrawParams

Withdraw params.

WithdrawReport

Report of the execution of withdrawal.

WithdrawalActionParams

Withdrawal params.

WithdrawalTokenAccounts

Token Accounts.

Enums

ActionState

Action State.

DecreasePositionSwapType

Swap Type for the decrease position action.

GlvPricingKind

Pricing kind.

GtUpdateKind

GT Update Kind.

InsolventCloseStep

Insolvent Close Step.

OrderKind

Order Kind.

PoolKind

Pool kind.